Moving static Block.fromJson(String type, JSONObject) to BlockFactory member method fromJson(String type, JSONObject).
Moving static Block.fromXml(XmlPullParser, BlockFactory)
to BlockFactory member method fromXml(XmlPullParser).

As member methods, this will allow BlockFactory to manage shared block
state data (e.g., forthcoming dropdown option lists).

Moved related tests from BlockTest to BlockFactoryTest,
and cleaning up along the way.
