+++ blocklylib-core/src/main/java/com/google/blockly/model/Block.java
@@ -649,18 +649,26 @@ public class Block {
     }
 
     /**
-     * @return The next or input {@link Connection} this block is connected to, or null.
+     * @return The previous or output {@link Connection} on this block, or null.
      */
-    public Connection getParentConnection() {
-        Connection prev = mPreviousConnection;
-        if (prev != null) {
-            return prev.getTargetConnection();
+    @Nullable
+    public Connection getUpwardsConnection() {
+        if (mPreviousConnection != null) {
+            return mPreviousConnection;
         }
-        Connection output = mOutputConnection;
-        if (output != null) {
-            return output.getTargetConnection();
+        if (mOutputConnection != null) {
+            return mOutputConnection;
         }
         return null;
+    }
+
+    /**
+     * @return The next or input {@link Connection} this block is connected to, or null.
+     */
+    @Nullable
+    public Connection getParentConnection() {
+        Connection upwards = getUpwardsConnection();
+        return upwards == null ? null : upwards.getTargetConnection();
     }
 
     /**
@@ -672,8 +680,42 @@ public class Block {
         return parentConnection == null ? null : parentConnection.getBlock();
     }
 
-    @VisibleForTesting
+    /**
+     * Configures whether this block should be a shadow block. This should only be called during
+     * creation of the block.
+     *
+     * @param isShadow
+     */
     void setShadow(boolean isShadow) {
+        if (mIsShadow == isShadow) {
+            return;
+        }
+
+        // Unsupported: Changing while connected to the parent (Connection target type, etc.)
+        Connection upConnection = getUpwardsConnection();
+        if (upConnection != null && upConnection.getTargetConnection() != null) {
+            throw new IllegalStateException(
+                    "Cannot change block shadow state while connected to parent.");
+        }
+        if (isShadow) {
+            if (containsVariableField(mInputList)) {
+                throw new IllegalStateException("Shadow blocks cannot contain variable fields.");
+            }
+
+            // Shadow blocks cannot have non-shadow children.
+            int inputCount = mInputList == null ? 0 : mInputList.size();
+            for (int i = 0; i < inputCount; ++i) {
+                Input input = mInputList.get(i);
+                Block child = input.getConnectedBlock();  // Not the shadow connection.
+                if (child != null && !child.isShadow()) {
+                    throw new IllegalStateException(
+                            "Cannot change block to shadow while non-shadow children are connected."
+                    );
+                }
+            }
+        }
+
+        // State change is valid. Proceed.
         mIsShadow = isShadow;
     }
 
@@ -787,379 +829,21 @@ public class Block {
     }
 
     /**
-     * Generate a Blockly Block from JSON. If the JSON is misformatted a {@link RuntimeException}
-     * will be thrown. All inputs and fields for the block will also be generated.
-     *
-     * @param type The type of the block.
-     * @param json The JSON to generate the block from.
-     *
-     * @return The generated Block.
+     * @return True if any input in this block includes a variable field. Otherwise false.
      */
-    public static Block fromJson(String type, JSONObject json) throws BlockLoadingException {
-        if (TextUtils.isEmpty(type)) {
-            throw new IllegalArgumentException("Block type may not be null or empty.");
-        }
-        if (json == null) {
-            throw new IllegalArgumentException("Json may not be null.");
-        }
-        Builder bob = new Builder(type);
-
-        if (json.has("output") && json.has("previousStatement")) {
-            throw new IllegalArgumentException(
-                    "Block cannot have both an output and a previous statement.");
-        }
-
-        // Parse any connections that are present.
-        if (json.has("output")) {
-            String[] checks = Input.getChecksFromJson(json, "output");
-            Connection output = new Connection(Connection.CONNECTION_TYPE_OUTPUT, checks);
-            bob.setOutput(output);
-        } else if (json.has("previousStatement")) {
-            String[] checks = Input.getChecksFromJson(json, "previousStatement");
-            Connection previous = new Connection(Connection.CONNECTION_TYPE_PREVIOUS, checks);
-            bob.setPrevious(previous);
-        }
-        // A block can have either an output connection or previous connection, but it can always
-        // have a next connection.
-        if (json.has("nextStatement")) {
-            String[] checks = Input.getChecksFromJson(json, "nextStatement");
-            Connection next = new Connection(Connection.CONNECTION_TYPE_NEXT, checks);
-            bob.setNext(next);
-        }
-        if (json.has("inputsInline")) {
-            try {
-                bob.setInputsInline(json.getBoolean("inputsInline"));
-            } catch (JSONException e) {
-                // Do nothing and it will remain false.
-            }
-        }
-
-        int blockColor = ColorUtils.DEFAULT_BLOCK_COLOR;
-        if (json.has("colour")) {
-            try {
-                String colourString = json.getString("colour");
-                blockColor = ColorUtils.parseColor(colourString, TEMP_IO_THREAD_FLOAT_ARRAY,
-                        ColorUtils.DEFAULT_BLOCK_COLOR);
-            } catch (JSONException e) {
-                // Won't get here. Checked above.
-            }
-        }
-        bob.setColor(blockColor);
-
-        ArrayList<Input> inputs = new ArrayList<>();
-        ArrayList<Field> fields = new ArrayList<>();
-        for (int i = 0; ; i++) {
-            String messageKey = "message" + i;
-            String argsKey = "args" + i;
-            String lastDummyAlignKey = "lastDummyAlign" + i;
-            if (!json.has(messageKey)) {
-                break;
-            }
-            String message = json.optString(messageKey);
-            JSONArray args = json.optJSONArray(argsKey);
-            if (args == null) {
-                // If there's no args for this message use an empty array.
-                args = new JSONArray();
-            }
-
-            if (message.matches("^%[a-zA-Z][a-zA-Z_0-9]*$")) {
-                // TODO(#83): load the message from resources.
-            }
-            // Split on all argument indices of the form "%N" where N is a number from 1 to
-            // the number of args without removing them.
-            List<String> tokens = tokenizeMessage(message);
-            int indexCount = 0;
-            // Indices start at 1, make the array 1 bigger so we don't have to offset things
-            boolean[] seenIndices = new boolean[args.length() + 1];
-
-            for (String token : tokens) {
-                // Check if this token is an argument index of the form "%N"
-                if (token.matches("^%\\d+$")) {
-                    int index = Integer.parseInt(token.substring(1));
-                    if (index < 1 || index > args.length()) {
-                        throw new IllegalArgumentException("Message index " + index
-                                + " is out of range.");
-                    }
-                    if (seenIndices[index]) {
-                        throw new IllegalArgumentException(("Message index " + index
-                                + " is duplicated"));
-                    }
-                    seenIndices[index] = true;
-
-                    JSONObject element;
-                    try {
-                        element = args.getJSONObject(index - 1);
-                    } catch (JSONException e) {
-                        throw new RuntimeException("Error reading arg %" + index, e);
-                    }
-                    while (element != null) {
-                        String elementType = element.optString("type");
-                        if (TextUtils.isEmpty(elementType)) {
-                            throw new IllegalArgumentException("No type for arg %" + index);
-                        }
-
-                        if (Field.isFieldType(elementType)) {
-                            fields.add(Field.fromJson(element));
-                            break;
-                        } else if (Input.isInputType(elementType)) {
-                            Input input = Input.fromJson(element);
-                            input.addAll(fields);
-                            fields.clear();
-                            inputs.add(input);
-                            break;
-                        } else {
-                            // Try getting the fallback block if it exists
-                            Log.w(TAG, "Unknown element type: " + elementType);
-                            element = element.optJSONObject("alt");
-                        }
-                    }
-                } else {
-                    token = token.replace("%%", "%").trim();
-                    if (!TextUtils.isEmpty(token)) {
-                        fields.add(new FieldLabel(null, token));
-                    }
-                }
-            }
-
-            // Verify every argument was used
-            for (int j = 1; j < seenIndices.length; j++) {
-                if (!seenIndices[j]) {
-                    throw new IllegalArgumentException("Argument " + j + " was never used.");
-                }
-            }
-            // If there were leftover fields we need to add a dummy input to hold them.
-            if (fields.size() != 0) {
-                String align = json.optString(lastDummyAlignKey, Input.ALIGN_LEFT_STRING);
-                Input input = new Input.InputDummy(null, align);
-                input.addAll(fields);
-                inputs.add(input);
-                fields.clear();
-            }
-        }
-
-        bob.setInputs(inputs);
-        return bob.build();
+    boolean containsVariableField() {
+        return containsVariableField(mInputList);
     }
 
     /**
-     * Load a block and all of its children from XML.
-     *
-     * @param parser An XmlPullParser pointed at the start tag of this block.
-     * @param factory A BlockFactory that will provide Blocks by name.
-     *
-     * @return The loaded block.
-     *
-     * @throws XmlPullParserException
-     * @throws IOException
-     * @throws BlocklyParserException
+     * @param inputs {@link Input}s to check for variable fields.
+     * @return True if any input in {@code inputs} includes a variable field. Otherwise false.
      */
-    public static Block fromXml(XmlPullParser parser, BlockFactory factory)
-            throws XmlPullParserException, IOException, BlocklyParserException {
-        String type = parser.getAttributeValue(null, "type");   // prototype name
-        String id = parser.getAttributeValue(null, "id");
-        if (type == null || type.isEmpty()) {
-            // If the id was empty the blockfactory will just generate one.
-            throw new BlocklyParserException("Block was missing a type.");
-        }
-
-        Block resultBlock = factory.obtainBlock(type, id);
-        if (resultBlock == null) {
-            throw new BlocklyParserException("Tried to obtain a block of an unknown type " + type);
-        }
-
-        String collapsedString = parser.getAttributeValue(null, "collapsed");
-        if (collapsedString != null) {
-            resultBlock.setCollapsed(Boolean.parseBoolean(collapsedString));
-        }
-
-        String deletableString = parser.getAttributeValue(null, "deletable");
-        if (deletableString != null) {
-            resultBlock.setDeletable(Boolean.parseBoolean(deletableString));
-        }
-
-        String disabledString = parser.getAttributeValue(null, "disabled");
-        if (disabledString != null) {
-            resultBlock.setDisabled(Boolean.parseBoolean(disabledString));
-        }
-
-        String editableString = parser.getAttributeValue(null, "editable");
-        if (editableString != null) {
-            resultBlock.setEditable(Boolean.parseBoolean(editableString));
-        }
-
-        String inputsInlineString = parser.getAttributeValue(null, "inline");
-        if (inputsInlineString != null) {
-            resultBlock.setInputsInline(Boolean.parseBoolean(inputsInlineString));
-        }
-
-        String movableString = parser.getAttributeValue(null, "movable");
-        if (movableString != null) {
-            resultBlock.setMovable(Boolean.parseBoolean(movableString));
-        }
-
-        // Set position.  Only if this is a top level block.
-        String x = parser.getAttributeValue(null, "x");
-        String y = parser.getAttributeValue(null, "y");
-        if (x != null && y != null) {
-            resultBlock.setPosition(Integer.parseInt(x), Integer.parseInt(y));
-        }
-
-        int eventType = parser.next();
-        String text = "";
-        String fieldName = "";
-        Block childBlock = null;
-        Block childShadow = null;
-        Input valueInput = null;
-        Input statementInput = null;
-
-        while (eventType != XmlPullParser.END_DOCUMENT) {
-            String tagname = parser.getName();
-            switch (eventType) {
-                case XmlPullParser.START_TAG:
-                    text = ""; // Ignore text from parent (or prior) block.
-                    if (tagname.equalsIgnoreCase("block")) {
-                        childBlock = fromXml(parser, factory);
-                    } else if (tagname.equalsIgnoreCase("shadow")) {
-                        childShadow = fromXml(parser, factory);
-                    } else if (tagname.equalsIgnoreCase("field")) {
-                        fieldName = parser.getAttributeValue(null, "name");
-                    } else if (tagname.equalsIgnoreCase("value")) {
-                        valueInput = resultBlock.getInputByName(
-                                parser.getAttributeValue(null, "name"));
-                        if (valueInput == null) {
-                            throw new BlocklyParserException("The value input was null at line "
-                                    + parser.getLineNumber() + "!");
-                        }
-                    } else if (tagname.equalsIgnoreCase("statement")) {
-                        statementInput = resultBlock.getInputByName(
-                                parser.getAttributeValue(null, "name"));
-                    } else if (tagname.equalsIgnoreCase("mutation")) {
-                        // TODO(fenichel): Handle mutations.
-                    }
-                    break;
-
-                case XmlPullParser.TEXT:
-                    text = parser.getText();
-                    break;
-
-                case XmlPullParser.END_TAG:
-                    Connection parentConnection = null;
-
-                    if (tagname.equalsIgnoreCase("block")) {
-                        if (resultBlock == null) {
-                            throw new BlocklyParserException(
-                                    "Created a null block. This should never happen.");
-                        }
-                        return resultBlock;
-                    } else if (tagname.equalsIgnoreCase("shadow")) {
-                        if (resultBlock == null) {
-                            throw new BlocklyParserException(
-                                    "Created a null block. This should never happen.");
-                        }
-                        if (containsVariableField(resultBlock.mInputList)) {
-                            throw new IllegalArgumentException(
-                                    "Shadow blocks cannot contain variable fields.");
-                        }
-                        resultBlock.mIsShadow = true;
-                        return resultBlock;
-                    }else if (tagname.equalsIgnoreCase("field")) {
-                        Field toSet = resultBlock.getFieldByName(fieldName);
-                        if (toSet != null) {
-                            if (!toSet.setFromString(text)) {
-                                throw new BlocklyParserException(
-                                        "Failed to set a field's value from XML.");
-                            }
-                        }
-                    } else if (tagname.equalsIgnoreCase("comment")) {
-                        resultBlock.setComment(text);
-                    } else if (tagname.equalsIgnoreCase("value")) {
-                        if (valueInput != null) {
-                            parentConnection = valueInput.getConnection();
-                            if (parentConnection == null) {
-                                throw new BlocklyParserException("The input connection was null.");
-                            }
-                        } else {
-                            throw new BlocklyParserException(
-                                    "A value input was null.");
-                        }
-                    } else if (tagname.equalsIgnoreCase("statement")) {
-                        if (statementInput != null) {
-                            parentConnection = statementInput.getConnection();
-                            if (parentConnection == null) {
-                                throw new BlocklyParserException(
-                                        "The statement connection was null.");
-                            }
-                        } else {
-                            throw new BlocklyParserException(
-                                    "A statement input was null.");
-                        }
-                    } else if (tagname.equalsIgnoreCase("next")) {
-                        parentConnection = resultBlock.getNextConnection();
-                        if (parentConnection == null) {
-                            throw new BlocklyParserException("A next connection was null");
-                        }
-                    }
-                    // If we finished a parent connection (statement, value, or next)
-                    if (parentConnection != null) {
-                        // Connect its child if one exists
-                        if (childBlock != null) {
-                            Connection childConnection = childBlock.getPreviousConnection();
-                            if (childConnection == null) {
-                                childConnection = childBlock.getOutputConnection();
-                            }
-                            if (childConnection == null) {
-                                throw new BlocklyParserException(
-                                        "The child block's connection was null.");
-                            }
-                            if (parentConnection.isConnected()) {
-                                throw new BlocklyParserException("Duplicated " + tagname
-                                        + " in block.");
-                            }
-                            parentConnection.connect(childConnection);
-                        }
-                        // Then connect its shadow if one exists
-                        if (childShadow != null) {
-                            Connection shadowConnection = childShadow.getPreviousConnection();
-                            if (shadowConnection == null) {
-                                shadowConnection = childShadow.getOutputConnection();
-                            }
-                            if (shadowConnection == null) {
-                                throw new BlocklyParserException(
-                                        "The shadow block connection was null.");
-                            }
-                            if (parentConnection.getShadowConnection() != null) {
-                                throw new BlocklyParserException("Duplicated " + tagname
-                                        + " in block.");
-                            }
-                            parentConnection.setShadowConnection(shadowConnection);
-                            if (!parentConnection.isConnected()) {
-                                // If there was no standard block connect the shadow
-                                parentConnection.connect(shadowConnection);
-                            }
-                        }
-                        // And clear out all the references for this tag
-                        childBlock = null;
-                        childShadow = null;
-                        valueInput = null;
-                        statementInput = null;
-                    }
-                    break;
-
-                default:
-                    break;
-            }
-            eventType = parser.next();
-        }
-        // Should never reach here, since this is called from a workspace fromXml function.
-        throw new BlocklyParserException(
-                "Reached the end of Block.fromXml. This should never happen.");
-    }
-
-    private static boolean containsVariableField(ArrayList<Input> mInputs) {
+    private static boolean containsVariableField(ArrayList<Input> inputs) {
         // Verify there's not a variable field
-        int inputCount = mInputs.size();
+        int inputCount = inputs.size();
         for (int i = 0; i < inputCount; i++) {
-            Input in = mInputs.get(i);
+            Input in = inputs.get(i);
             List<Field> fields = in.getFields();
             int fieldCount = fields.size();
             for (int j = 0; j < fieldCount; j++) {
