+++ blocklylib-core/src/main/java/com/google/blockly/model/ToolboxCategory.java
@@ -152,7 +152,7 @@ public class ToolboxCategory {
                     if (parser.getName().equalsIgnoreCase("category")) {
                         result.addSubcategory(ToolboxCategory.fromXml(parser, factory));
                     } else if (parser.getName().equalsIgnoreCase("block")) {
-                        result.addBlock(Block.fromXml(parser, factory));
+                        result.addBlock(factory.fromXml(parser));
                     } else if (parser.getName().equalsIgnoreCase("shadow")) {
                         throw new IllegalArgumentException(
                                 "Shadow blocks may not be top level toolbox blocks.");
