+++ blocklytest/src/androidTest/java/com/google/blockly/model/BlockFactoryTest.java
@@ -17,29 +17,213 @@ package com.google.blockly.model;
 import android.test.AndroidTestCase;
 
 import com.google.blockly.android.R;
+import com.google.blockly.utils.BlockLoadingException;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.xmlpull.v1.XmlPullParser;
+import org.xmlpull.v1.XmlPullParserException;
+import org.xmlpull.v1.XmlPullParserFactory;
+
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.util.List;
+
+import static com.google.blockly.utils.MoreAsserts.assertStringNotEmpty;
+
 
 /**
  * Tests for {@link BlockFactory}.
  */
 public class BlockFactoryTest extends AndroidTestCase {
+    private XmlPullParserFactory xmlPullParserFactory;
+    private BlockFactory mBlockFactory;
 
-    // TODO(#84): Move test_blocks.json to the testapp's resources.
+    @Override
+    protected void setUp() throws Exception {
+        super.setUp();
+        xmlPullParserFactory = XmlPullParserFactory.newInstance();
+        // TODO(#84): Move test_blocks.json to the testapp's resources.
+        mBlockFactory = new BlockFactory(getContext(), new int[]{R.raw.test_blocks});
+    }
+
+    public void testJson() throws JSONException, BlockLoadingException {
+        JSONObject blockDefinitionJson = new JSONObject(BlockTestStrings.TEST_JSON_STRING);
+        Block block = mBlockFactory.fromJson("test_block", blockDefinitionJson);
+
+        assertNotNull("Block was null after initializing from JSON", block);
+        assertEquals("Type not set correctly", "test_block", block.getType());
+        assertStringNotEmpty("Block id cannot be empty.", block.getId());
+        assertEquals("Wrong number of inputs", 2, block.getInputs().size());
+        assertEquals("Wrong number of fields in first input",
+                9, block.getInputs().get(0).getFields().size());
+    }
+
 
     public void testLoadBlocks() {
-        BlockFactory bf = new BlockFactory(getContext(), new int[] {R.raw.test_blocks});
-        List<Block> blocks = bf.getAllBlocks();
+        List<Block> blocks = mBlockFactory.getAllBlocks();
         assertEquals("BlockFactory failed to load all blocks.", 20, blocks.size());
     }
 
+    public void testLoadFromXml() throws IOException, XmlPullParserException {
+        Block loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK);
+        assertEquals("frankenblock", loaded.getType());
+        assertEquals(new WorkspacePoint(37, 13), loaded.getPosition());
+
+        // All blocks need a type.
+        parseBlockFromXmlFailure(BlockTestStrings.NO_BLOCK_TYPE,
+                "Block without a type must fail to load.");
+
+        // Ids can be generated by the BlockFactory.
+        parseBlockFromXml(BlockTestStrings.NO_BLOCK_ID);
+
+        // Only top level blocks need a position.
+        parseBlockFromXml(BlockTestStrings.NO_BLOCK_POSITION);
+
+        // Values.
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "1",
+                BlockTestStrings.VALUE_GOOD));
+        // Value: null child block
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "3",
+                BlockTestStrings.VALUE_NO_CHILD));
+
+        // TODO(fenichel): Value: no input connection
+
+        // Value: no output connection on child
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "2",
+                BlockTestStrings.VALUE_NO_OUTPUT),
+                "A <value> child block without an output must fail to load.");
+
+        // Value: no input with that name
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "4",
+                BlockTestStrings.VALUE_BAD_NAME),
+                "A block without a recognized type id must fail to load.");
+        // Value: multiple values for the same input
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "5",
+                BlockTestStrings.VALUE_REPEATED),
+                "An input <value> with multiple blocks must fail to load.");
+
+        // Comment: with text
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "6",
+                BlockTestStrings.COMMENT_GOOD));
+        // Comment: empty string
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "7",
+                BlockTestStrings.COMMENT_NO_TEXT));
+
+        // Fields
+        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "8",
+                BlockTestStrings.FIELD_HAS_NAME));
+        assertEquals("item", ((FieldInput) loaded.getFieldByName("text_input")).getText());
+        // A missing or unknown field name isn't an error, it's just ignored.
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "9",
+                BlockTestStrings.FIELD_MISSING_NAME));
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "10",
+                BlockTestStrings.FIELD_UNKNOWN_NAME));
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "11",
+                BlockTestStrings.FIELD_MISSING_TEXT));
+
+        // Statement: no previous connection on child block
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "13",
+                BlockTestStrings.STATEMENT_BAD_CHILD),
+                "A statement <value> child without a previous connection must fail to load.");
+        // Statement: no input with that name
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "14",
+                BlockTestStrings.STATEMENT_BAD_NAME),
+                "A block without a recognized type id must fail to load.");
+
+        // Statement
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "15",
+                BlockTestStrings.STATEMENT_GOOD));
+        // Statement: null child block
+        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "12",
+                BlockTestStrings.STATEMENT_NO_CHILD));
+    }
+
+    public void testLoadFromXmlInline() throws IOException, XmlPullParserException {
+        Block loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK_INLINE_BEGINNING);
+        assertTrue(loaded.getInputsInline());
+        assertTrue(loaded.getInputsInlineModified());
+
+        loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK_INLINE_END);
+        assertTrue(loaded.getInputsInline());
+        assertTrue(loaded.getInputsInlineModified());
+
+        loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK_INLINE_FALSE);
+        assertFalse(loaded.getInputsInline());
+        assertTrue(loaded.getInputsInlineModified());
+
+        loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK);
+        assertFalse(loaded.getInputsInline());
+        assertFalse(loaded.getInputsInlineModified());
+    }
+
+    public void testLoadFromXml_shadows() throws IOException, XmlPullParserException {
+        Block loaded = parseShadowFromXml(BlockTestStrings.SIMPLE_SHADOW);
+        assertEquals("math_number", loaded.getType());
+        assertEquals(new WorkspacePoint(37, 13), loaded.getPosition());
+        assertTrue(loaded.isShadow());
+
+
+        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "1",
+                BlockTestStrings.VALUE_SHADOW));
+        Connection conn = loaded.getInputByName("value_input").getConnection();
+        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
+        assertTrue(conn.getShadowBlock().isShadow());
+
+        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "2",
+                BlockTestStrings.VALUE_SHADOW_GOOD));
+        conn = loaded.getInputByName("value_input").getConnection();
+        assertEquals("VALUE_REAL", conn.getTargetBlock().getId());
+        assertFalse(conn.getTargetBlock().isShadow());
+        assertEquals("VALUE_SHADOW", conn.getShadowBlock().getId());
+        assertTrue(conn.getShadowBlock().isShadow());
+
+        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "3",
+                BlockTestStrings.STATEMENT_SHADOW));
+        conn = loaded.getInputByName("NAME").getConnection();
+        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
+        assertTrue(conn.getShadowBlock().isShadow());
+
+        // Clear refs so block names can be reused
+        mBlockFactory.clearPriorBlockReferences();
+        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "4",
+                BlockTestStrings.STATEMENT_SHADOW_GOOD));
+        conn = loaded.getInputByName("NAME").getConnection();
+        assertEquals("STATEMENT_REAL", conn.getTargetBlock().getId());
+        assertFalse(conn.getTargetBlock().isShadow());
+        assertEquals("STATEMENT_SHADOW", conn.getShadowBlock().getId());
+        assertTrue(conn.getShadowBlock().isShadow());
+
+        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "5",
+                BlockTestStrings.VALUE_NESTED_SHADOW));
+        conn = loaded.getInputByName("value_input").getConnection();
+        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
+        Block shadow1 = conn.getShadowBlock();
+        assertEquals("SHADOW1", shadow1.getId());
+        conn = shadow1.getOnlyValueInput().getConnection();
+        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
+        assertEquals("SHADOW2", conn.getShadowBlock().getId());
+
+        // Clear refs so block names can be reused
+        mBlockFactory.clearPriorBlockReferences();
+
+        // Verify a shadow block with a shadow child fails.
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "6",
+                BlockTestStrings.VALUE_NESTED_SHADOW_BLOCK),
+                "A <shadow> containing a normal <block> must fail to load.");
+
+        // Verify a block with a variable cannot be turned into a shadow block
+        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "7",
+                BlockTestStrings.VALUE_SHADOW_VARIABLE),
+                "A <shadow> containing a variable field (at any depth) must fail to load.");
+    }
+
     public void testObtainBlock() {
-        BlockFactory bf = new BlockFactory(getContext(), new int[] {R.raw.test_blocks});
-        Block emptyBlock = bf.obtainBlock("empty_block", null);
+        Block emptyBlock = mBlockFactory.obtainBlock("empty_block", null);
         assertNotNull("Failed to create the empty block.", emptyBlock);
         assertEquals("Empty block has the wrong type", "empty_block", emptyBlock.getType());
 
-        Block frankenblock = bf.obtainBlock("frankenblock", null);
+        Block frankenblock = mBlockFactory.obtainBlock("frankenblock", null);
         assertNotNull("Failed to create the frankenblock.", frankenblock);
 
         List<Input> inputs = frankenblock.getInputs();
@@ -55,11 +239,10 @@ public class BlockFactoryTest extends AndroidTestCase {
     }
 
     public void testObtainBlock_repeatedWithoutUuid() {
-        BlockFactory bf = new BlockFactory(getContext(), new int[] {R.raw.test_blocks});
-        Block frankenblock = bf.obtainBlock("frankenblock", null);
+        Block frankenblock = mBlockFactory.obtainBlock("frankenblock", null);
         assertNotNull("Failed to create the frankenblock.", frankenblock);
 
-        Block frankencopy = bf.obtainBlock("frankenblock", null);
+        Block frankencopy = mBlockFactory.obtainBlock("frankenblock", null);
         assertNotSame("Obtained blocks should be distinct objects when uuid is null.",
                 frankenblock, frankencopy);
 
@@ -73,12 +256,11 @@ public class BlockFactoryTest extends AndroidTestCase {
     }
 
     public void testObtainBlock_repeatedWithUuid() {
-        BlockFactory bf = new BlockFactory(getContext(), new int[] {R.raw.test_blocks});
-        Block frankenblock = bf.obtainBlock("frankenblock", "123");
+        Block frankenblock = mBlockFactory.obtainBlock("frankenblock", "123");
         assertNotNull("Failed to create the frankenblock.", frankenblock);
 
         try {
-            Block frankencopy = bf.obtainBlock("frankenblock", "123");
+            mBlockFactory.obtainBlock("frankenblock", "123");
             fail("Cannot create two blocks with the same id");
         } catch (IllegalArgumentException e) {
             // expected
@@ -86,11 +268,11 @@ public class BlockFactoryTest extends AndroidTestCase {
     }
 
     public void testObtainBlock_repeatedWithUuidMismatchingPrototype() {
-        BlockFactory bf = new BlockFactory(getContext(), new int[] {R.raw.test_blocks});
-        Block frankenblock = bf.obtainBlock("frankenblock", "123");
+        Block frankenblock = mBlockFactory.obtainBlock("frankenblock", "123");
+        assertNotNull("Failed to create the frankenblock.", frankenblock);
 
         try {
-            bf.obtainBlock("empty_block", "123");
+            mBlockFactory.obtainBlock("empty_block", "123");
 
             // Should not get here.
             fail("Expected error when requesting a block with matching UUID "
@@ -98,5 +280,58 @@ public class BlockFactoryTest extends AndroidTestCase {
         } catch(IllegalArgumentException e) {
             // Expected.
         }
+    }
+
+    private Block parseBlockFromXml(String testString)
+            throws IOException, XmlPullParserException {
+        XmlPullParser parser = getXmlPullParser(testString, "block");
+        Block loaded = mBlockFactory.fromXml(parser);
+        assertNotNull(loaded);
+        return loaded;
+    }
+
+    private Block parseShadowFromXml(String testString)
+            throws IOException, XmlPullParserException {
+        XmlPullParser parser = getXmlPullParser(testString, "shadow");
+        Block loaded = mBlockFactory.fromXml(parser);
+        assertNotNull(loaded);
+        return loaded;
+    }
+
+    private void parseBlockFromXmlFailure(String testString, String messageIfDoesNotFail)
+            throws IOException, XmlPullParserException {
+
+        XmlPullParser parser = getXmlPullParser(testString, "block");
+        try {
+            mBlockFactory.fromXml(parser);
+            fail(messageIfDoesNotFail);
+        } catch (BlocklyParserException expected) {
+            // expected
+        }
+    }
+
+    /**
+     * Creates a pull parser with the given input and gobbles up to the first start tag that equals
+     * {@code returnFirstInstanceOf}.
+     */
+    private XmlPullParser getXmlPullParser(String input, String returnFirstInstanceOf) {
+        try {
+            xmlPullParserFactory.setNamespaceAware(true);
+            XmlPullParser parser = xmlPullParserFactory.newPullParser();
+
+            parser.setInput(new ByteArrayInputStream(input.getBytes()), null);
+
+            int eventType = parser.getEventType();
+            while (eventType != XmlPullParser.END_DOCUMENT) {
+                if (eventType == XmlPullParser.START_TAG
+                        && parser.getName().equalsIgnoreCase(returnFirstInstanceOf)) {
+                    return parser;
+                }
+                eventType = parser.next();
+            }
+        } catch (XmlPullParserException | IOException e) {
+            throw new BlocklyParserException(e);
+        }
+        return null;
     }
 }
