+++ blocklytest/src/androidTest/java/com/google/blockly/model/BlockTest.java
@@ -18,13 +18,9 @@ package com.google.blockly.model;
 import android.test.AndroidTestCase;
 
 import com.google.blockly.android.R;
-import com.google.blockly.utils.BlockLoadingException;
 import com.google.blockly.utils.BlocklyXmlHelper;
 import com.google.blockly.utils.StringOutputStream;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
 import org.xmlpull.v1.XmlPullParserFactory;
 import org.xmlpull.v1.XmlSerializer;
@@ -33,16 +29,13 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.StringReader;
 import java.io.UnsupportedEncodingException;
-import java.io.Writer;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import static android.test.MoreAsserts.assertNotEqual;
+import static com.google.blockly.utils.MoreAsserts.assertStringNotEmpty;
 
 /**
  * Tests for {@link Block}.
@@ -56,18 +49,6 @@ public class BlockTest extends AndroidTestCase {
         super.setUp();
         xmlPullParserFactory = XmlPullParserFactory.newInstance();
         mBlockFactory = new BlockFactory(getContext(), new int[]{R.raw.test_blocks});
-    }
-
-    public void testJson() throws JSONException, BlockLoadingException {
-        JSONObject blockDefinitionJson = new JSONObject(BlockTestStrings.TEST_JSON_STRING);
-        Block block = Block.fromJson("test_block", blockDefinitionJson);
-
-        assertNotNull("Block was null after initializing from JSON", block);
-        assertEquals("Type not set correctly", "test_block", block.getType());
-        assertStringNotEmpty("Block id cannot be empty.", block.getId());
-        assertEquals("Wrong number of inputs", 2, block.getInputs().size());
-        assertEquals("Wrong number of fields in first input",
-                9, block.getInputs().get(0).getFields().size());
     }
 
     public void testEmptyBlockHasId() {
@@ -131,163 +112,6 @@ public class BlockTest extends AndroidTestCase {
         tokens = Block.tokenizeMessage(testMessage);
         expected = Arrays.asList(new String[]{"%Hello", "%1", "World%"});
         assertListsMatch(expected, tokens);
-    }
-
-    public void testLoadFromXml() throws IOException, XmlPullParserException {
-        // TODO(#84): Move test_blocks.json to the test app's resources
-        BlockFactory bf = new BlockFactory(getContext(), new int[]{R.raw.test_blocks});
-
-        Block loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK, bf);
-        assertEquals("frankenblock", loaded.getType());
-        assertEquals(new WorkspacePoint(37, 13), loaded.getPosition());
-
-        // All blocks need a type. Ids can be generated by the BlockFactory.
-        parseBlockFromXmlFailure(BlockTestStrings.NO_BLOCK_TYPE, bf);
-        parseBlockFromXml(BlockTestStrings.NO_BLOCK_ID, bf);
-
-        // Only top level blocks need a position.
-        parseBlockFromXml(BlockTestStrings.NO_BLOCK_POSITION, bf);
-
-        // Values.
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "1",
-                BlockTestStrings.VALUE_GOOD), bf);
-        // Value: null child block
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "3",
-                BlockTestStrings.VALUE_NO_CHILD), bf);
-        // TODO(fenichel): Value: no input connection
-        // Value: no output connection on child
-        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "2",
-                BlockTestStrings.VALUE_NO_OUTPUT), bf);
-        // Value: no input with that name
-        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "4",
-                BlockTestStrings.VALUE_BAD_NAME), bf);
-        // Value: multiple values for the same input
-        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "5",
-                BlockTestStrings.VALUE_REPEATED), bf);
-
-        // Comment: with text
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "6",
-                BlockTestStrings.COMMENT_GOOD), bf);
-        // Comment: empty string
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "7",
-                BlockTestStrings.COMMENT_NO_TEXT), bf);
-
-        // Fields
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "8",
-                BlockTestStrings.FIELD_HAS_NAME), bf);
-        assertEquals("item", ((FieldInput) loaded.getFieldByName("text_input")).getText());
-        // A missing or unknown field name isn't an error, it's just ignored.
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "9",
-                BlockTestStrings.FIELD_MISSING_NAME), bf);
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "10",
-                BlockTestStrings.FIELD_UNKNOWN_NAME), bf);
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "11",
-                BlockTestStrings.FIELD_MISSING_TEXT), bf);
-
-        // Statement: no previous connection on child block
-        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "13",
-                BlockTestStrings.STATEMENT_BAD_CHILD), bf);
-        // Statement: no input with that name
-        parseBlockFromXmlFailure(BlockTestStrings.assembleFrankenblock("block", "14",
-                BlockTestStrings.STATEMENT_BAD_NAME), bf);
-
-        // Statement
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "15",
-                BlockTestStrings.STATEMENT_GOOD), bf);
-        // Statement: null child block
-        parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "12",
-                BlockTestStrings.STATEMENT_NO_CHILD), bf);
-    }
-
-    public void testLoadFromXmlInline() throws IOException, XmlPullParserException {
-        // TODO(#84): Move test_blocks.json to the test app's resources
-        BlockFactory bf = new BlockFactory(getContext(), new int[]{R.raw.test_blocks});
-
-        Block loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK_INLINE_BEGINNING, bf);
-        assertTrue(loaded.getInputsInline());
-        assertTrue(loaded.getInputsInlineModified());
-
-        loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK_INLINE_END, bf);
-        assertTrue(loaded.getInputsInline());
-        assertTrue(loaded.getInputsInlineModified());
-
-        loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK_INLINE_FALSE, bf);
-        assertFalse(loaded.getInputsInline());
-        assertTrue(loaded.getInputsInlineModified());
-
-        loaded = parseBlockFromXml(BlockTestStrings.SIMPLE_BLOCK, bf);
-        assertFalse(loaded.getInputsInline());
-        assertFalse(loaded.getInputsInlineModified());
-    }
-
-    public void testLoadFromXml_shadows() throws IOException, XmlPullParserException {
-        BlockFactory bf = new BlockFactory(getContext(), new int[]{R.raw.test_blocks});
-
-        Block loaded = parseShadowFromXml(BlockTestStrings.SIMPLE_SHADOW, bf);
-        assertEquals("math_number", loaded.getType());
-        assertEquals(new WorkspacePoint(37, 13), loaded.getPosition());
-        assertTrue(loaded.isShadow());
-
-
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "1",
-                BlockTestStrings.VALUE_SHADOW), bf);
-        Connection conn = loaded.getInputByName("value_input").getConnection();
-        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
-        assertTrue(conn.getShadowBlock().isShadow());
-
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "2",
-                BlockTestStrings.VALUE_SHADOW_GOOD), bf);
-        conn = loaded.getInputByName("value_input").getConnection();
-        assertEquals("VALUE_REAL", conn.getTargetBlock().getId());
-        assertFalse(conn.getTargetBlock().isShadow());
-        assertEquals("VALUE_SHADOW", conn.getShadowBlock().getId());
-        assertTrue(conn.getShadowBlock().isShadow());
-
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "3",
-                BlockTestStrings.STATEMENT_SHADOW), bf);
-        conn = loaded.getInputByName("NAME").getConnection();
-        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
-        assertTrue(conn.getShadowBlock().isShadow());
-
-        // Clear refs so block names can be reused
-        bf.clearPriorBlockReferences();
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "4",
-                BlockTestStrings.STATEMENT_SHADOW_GOOD), bf);
-        conn = loaded.getInputByName("NAME").getConnection();
-        assertEquals("STATEMENT_REAL", conn.getTargetBlock().getId());
-        assertFalse(conn.getTargetBlock().isShadow());
-        assertEquals("STATEMENT_SHADOW", conn.getShadowBlock().getId());
-        assertTrue(conn.getShadowBlock().isShadow());
-
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "5",
-                BlockTestStrings.VALUE_NESTED_SHADOW), bf);
-        conn = loaded.getInputByName("value_input").getConnection();
-        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
-        Block shadow1 = conn.getShadowBlock();
-        assertEquals("SHADOW1", shadow1.getId());
-        conn = shadow1.getOnlyValueInput().getConnection();
-        assertEquals(conn.getTargetBlock(), conn.getShadowBlock());
-        assertEquals("SHADOW2", conn.getShadowBlock().getId());
-
-        // Clear refs so block names can be reused
-        bf.clearPriorBlockReferences();
-        loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "6",
-                BlockTestStrings.VALUE_NESTED_SHADOW_BLOCK), bf);
-        conn = loaded.getInputByName("value_input").getConnection();
-        shadow1 = conn.getShadowBlock();
-        assertEquals("SHADOW1", shadow1.getId());
-        conn = shadow1.getOnlyValueInput().getConnection();
-        assertEquals("BLOCK_INNER", conn.getTargetBlock().getId());
-        assertEquals("SHADOW2", conn.getShadowBlock().getId());
-
-        // Verify a block with a variable cannot be turned into a shadow block
-        try {
-            loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "7",
-                    BlockTestStrings.VALUE_SHADOW_VARIABLE), bf);
-            fail("Shadow blocks may not contain variables.");
-        } catch (IllegalArgumentException e) {
-            // Expected
-        }
     }
 
     public void testSerializeBlock() throws BlocklySerializerException, IOException {
@@ -492,59 +316,6 @@ public class BlockTest extends AndroidTestCase {
         // Statement input, dummy input and value input.
         underTest = bf.obtainBlock("controls_repeat_ext", null);
         assertSame(underTest.getInputByName("TIMES"), underTest.getOnlyValueInput());
-    }
-
-    private Block parseBlockFromXml(String testString, BlockFactory bf)
-            throws IOException, XmlPullParserException {
-        XmlPullParser parser = getXmlPullParser(testString, "block");
-        Block loaded = Block.fromXml(parser, bf);
-        assertNotNull(loaded);
-        return loaded;
-    }
-
-    private Block parseShadowFromXml(String testString, BlockFactory bf)
-            throws IOException, XmlPullParserException {
-        XmlPullParser parser = getXmlPullParser(testString, "shadow");
-        Block loaded = Block.fromXml(parser, bf);
-        assertNotNull(loaded);
-        return loaded;
-    }
-
-    private void parseBlockFromXmlFailure(String testString, BlockFactory bf)
-            throws IOException, XmlPullParserException {
-
-        XmlPullParser parser = getXmlPullParser(testString, "block");
-        try {
-            Block.fromXml(parser, bf);
-            fail("Should have thrown a BlocklyParseException.");
-        } catch (BlocklyParserException expected) {
-            // expected
-        }
-    }
-
-    /* Creates a pull parser with the given input and gobbles up to the first start tag
-     * that equals returnFirstInstanceOf.
-     */
-    private XmlPullParser getXmlPullParser(String input, String returnFirstInstanceOf) {
-        XmlPullParser parser = null;
-        try {
-            xmlPullParserFactory.setNamespaceAware(true);
-            parser = xmlPullParserFactory.newPullParser();
-
-            parser.setInput(new ByteArrayInputStream(input.getBytes()), null);
-
-            int eventType = parser.getEventType();
-            while (eventType != XmlPullParser.END_DOCUMENT) {
-                if (eventType == XmlPullParser.START_TAG
-                        && parser.getName().equalsIgnoreCase(returnFirstInstanceOf)) {
-                    return parser;
-                }
-                eventType = parser.next();
-            }
-        } catch (XmlPullParserException | IOException e) {
-            throw new BlocklyParserException(e);
-        }
-        return null;
     }
 
     private XmlSerializer getXmlSerializer(ByteArrayOutputStream os) throws BlocklySerializerException {
@@ -815,14 +586,6 @@ public class BlockTest extends AndroidTestCase {
             return BlocklyXmlHelper.loadOneBlockFromXml(stream, mBlockFactory);
         } catch (UnsupportedEncodingException e) {
             throw new IllegalStateException(e);
-        }
-    }
-
-    private void assertStringNotEmpty(String mesg, String str) {
-        if (str == null) {
-            fail(mesg + " Found null string.");
-        } else if(str.length() == 0) {
-            fail(mesg + " Found empty string.");
         }
     }
 }
